Simple version of DSL
* DSL is wired directly to markdown rendering methods for simplicity
* Block elements now have a trailing newline because that seems to make more sense
* Basic document test

Right now the dsl delegation is making the link and image methods act weird because
they're being added like block elements which doesn't make sense. Not sure whether
it's better to separate rendering and dsl, or change the dsl so that the
"add this to the document" part is left to the user e.g.

MDGen.md do |md|
  md << "Raw text"
  md << link 'whatever', 'http://example.com'
  md << "raw text after link
end